client/requirements.txt
grpcio
grpcio-tools
yfinance